DataIngestion

Database interface

DataIngestion.Repository — Type

Repository(db::DuckDB.DB)

Construct a Repository object that holds a DuckDB.DB as well as a pool of connections.

Use DBInterface.(f::Base.Callable, repo::Repository, sql::AbstractString, [params]) to run a function on the result of a query sql on an available connection in the pool.

Ingestion interface

DataIngestion.is_supported — Function

is_supported(file::AbstractString)

Denote whether a file can be read automatically.

DataIngestion.load_files — Function

load_files(
    repository::Repository, files::AbstractVector{<:AbstractString},
    [format::AbstractString]
)

Load files into a table called TABLE_NAMES.source inside repository.db. The format is inferred or can be passed explicitly.

Supported formats are

"csv",

"tsv",

"txt",

"json",

"parquet".

Selection interface

DataIngestion.Filters — Type

struct Filters
    filters::Vector{AbstractFilter}
end

Container for a list of filters.

DataIngestion.select — Function

select(filters::Filters, repo::Repository)

Create a table with name TABLE_NAMES.selection within the database repo.db based on rows from the table TABLE_NAMES.source that are kept by the filters in filters.

See also Filters and Repository.

Filters

DataIngestion.IntervalFilter — Type

struct IntervalFilter{T} <: AbstractFilter
    colname::String
    interval::ClosedInterval{T}
end

Object to retain only those rows for which the variable colname lies inside the interval.

DataIngestion.ListFilter — Type

struct ListFilter{T} <: AbstractFilter
    colname::String
    list::Vector{T}
end

Object to retain only those rows for which the variable colname belongs to a list of options.

Metadata for filter generation

DataIngestion.summarize — Function

summarize(repo::Repository, tbl::AbstractString)

Compute summaries of variables in table tbl within the database repo.db. The summary of a variable depends on its type, according to the following rules.

Categorical variable => list of unique types.

Continuous variable => extrema.