Pipelines

Transformation interface

Pipelines.Cards — Type

struct Cards
    cards::Vector{AbstractCard}
end

Container for a list of cards.

Pipelines.evaluate — Function

evaluate(cards::Cards, repo::Repository, table::AbstractString)

Replace table in the database repo.db with the outcome of executing all the transformations in cards.

Cards

Pipelines.RescaleCard — Type

struct RescaleCard <: AbstractCard
    method::String
    by::Vector{String}
    columns::Vector{String}
    suffix::String
end

Card to rescale of one or more columns according to a given method. The supported methods are

"zscore",

"maxabs",

"minmax",

"log",

"logistic".

The resulting rescaled variable is added to the table under the name "$(originalname)_$(suffix)".

Pipelines.SplitCard — Type

struct SplitCard <: AbstractCard
    method::String
    order_by::Vector{String}
    by::Vector{String}
    output::String
    p::Float64
    tiles::Vector{Int}
end

Card to split the data into two groups according to a given method.

Currently supported methods are

"tiles" (requires tiles argument, e.g., tiles = [1, 1, 2, 1, 1, 2]),

"percentile" (requires p argument, e.g. p = 0.9).