DuckDBUtils

Database interface

DuckDBUtils.Repository — Type

Repository(db::DuckDB.DB)

Construct a Repository object that holds a DuckDB.DB as well as a pool of connections.

Use DBInterface.(f::Base.Callable, repo::Repository, sql::AbstractString, [params]) to run a function on the result of a query sql on an available connection in the pool.

DuckDBUtils.get_catalog — Function

get_catalog(repo::Repository; schema = nothing)

Extract the catalog of available tables from a Repository repo.

DuckDBUtils.acquire_connection — Function

acquire_connection(repo::Repository)

Acquire an open connection to the database repo.db from the pool repo.pool. See also release_connection.

Note

A command con = acquire_connection(repo) must always be followed by a matching command release_connection(repo, con) (after the connection has been used).

DuckDBUtils.release_connection — Function

release_connection(repo::Repository, con)

Release connection con to the pool repo.pool

DuckDBUtils.with_connection — Function

with_connection(f, repo::Repository, [N])

Acquire a connection con from the pool repo.pool. Then, execute f(con) and release the connection to the pool. An optional parameter N can be passed to determine the number of connections to be acquired (defaults to 1).

DuckDBUtils.with_table — Function

with_table(f, repo::Repository, table)

Register a table under a random unique name name, apply f(name), and then unregister the table.

Batched iteration

DuckDBUtils.Batches — Type

struct Batches{T}
    chunks::T
    batchsize::Int
    nrows::Int
end

Let chunks be a partitioned table with nrows in total. Then, return an iterator of column-based tables with batchsize rows each.

Note

chunks can in general be obtained as the output of Tables.partitions.

Internal functions

DuckDBUtils._numobs — Function

_numobs(cols)

Compute the number of rows of a column-based table cols.

DuckDBUtils._init — Function

_init(cols)

Initialize an empty table with the same schema as the column-based table cols.

DuckDBUtils._append! — Function

_append!(batch::AbstractDict, cols, rg = Colon())

Append rows rg of column-based table cols to the dict table batch.