StreamlinerCore

StreamlinerCore is a julia library to generate, train and evaluate models defined via some configuration files.

Data interface

StreamlinerCore.AbstractData — Type

AbstactData{N}

Abstract type representing streamers of N datasets. In general, StreamlinerCore will use N = 1 to validate and evaluate trained models and N = 2 to train models via a training and a validation datasets.

Subtypes of AbstractData are meant to implement the following methods:

stream,

get_templates,

get_metadata,

get_summary (optional).

StreamlinerCore.stream — Function

stream(f, data::AbstractData{N}, partition::Integer; batchsize, device, rng::RNG = get_rng(), shuffle = false) where {N}

Stream partition of data by batches of batchsize on a given device. Return the result of applying f on the resulting batch iterator. Shuffling is optional and controlled by shuffle (boolean) and by the random number generator rng.

StreamlinerCore.ingest — Function

ingest(data::AbstractData{1}, eval_stream, select)

Ingest output of evaluate into a suitable database, tensor or iterator. select determines which fields of the model output to keep.

StreamlinerCore.get_templates — Function

get_templates(data::AbstractData)

Extract templates for data. Templates encode type and size of the arrays that data will stream. See also Template

StreamlinerCore.get_metadata — Function

get_metadata(x)::Dict{String, Any}

Extract metadata for x. metadata should be a dictionary of information that identifies x univoquely. After training, it will be stored in the MongoDB together. get_metadata has methods for AbstractData, Model, and Training.

StreamlinerCore.get_summary — Function

get_summary(data::AbstractData)::Dict{String, Any}

Extract summary for data. summary should be a dictionary of summary statistics for data. Common choices of statistics to report are mean and standard deviation, as well as unique values for categorical variables.

StreamlinerCore.get_nsamples — Function

get_nsamples(data::AbstractData{N})::NTuple{N, Int} where {N}

Return number of samples for data.

StreamlinerCore.Template — Type

Template(::Type{T}, size::NTuple{N, Int}) where {T, N}

Create an object of type Template. It represents arrays with eltype T and size size. Note that size does not include the minibatch dimension.

Parser

StreamlinerCore.Parser — Type

Parser(;
    model, layers, sigmas, aggregators, metrics, regularizations,
    optimizers, schedules, stoppers, devices
)

Collection of dictionaries to performance the necessary conversion from the user-specified configuration file or dictionary to julia objects.

For most usecases, one should define a default parser

parser = default_parser()

and pass it to Model and Training upon construction.

A parser object is also required to use interface functions that read from the MongoDB:

finetune,

loadmodel,

validate,

evaluate.

See default_parser for more advanced uses.

StreamlinerCore.default_parser — Function

default_parser(; plugins::AbstractVector{Parser}=Parser[])

Return a parser::Parser object that includes StreamlinerCore defaults together with optional plugins.

Parsed objects

StreamlinerCore.Model — Type

Model(parser::Parser, metadata::AbstractDict)

Create a Model object from a configuration dictionary metadata. The parser::Parser handles conversion from configuration variables to julia objects.

Given a model::Model object, use model(get_templates(data)) where data::AbstractData to instantiate the corresponding neural network or machine.

Model(parser::Parser, path::AbstractString, [vars::AbstractDict])

Create a Model object from a configuration dictionary stored at path in TOML format.

The parser::Parser handles conversion from configuration variables to julia objects.

The optional argument vars is a dictionary of variables the can be used to fill the template given in path.

See the static/model folder for example configurations.

StreamlinerCore.Training — Type

Training(parser::Parser, metadata::AbstractDict)

Create a Training object from a configuration dictionary metadata. The parser::Parser handles conversion from configuration variables to julia objects.

Training(parser::Parser, path::AbstractString, [vars::AbstractDict])

Create a Training object from a configuration dictionary stored at path in TOML format.

The parser::Parser handles conversion from configuration variables to julia objects.

The optional argument vars is a dictionary of variables the can be used to fill the template given in path.

See the static/training folder for example configurations.

Training and evaluation

StreamlinerCore.train — Function

train(
    model::Model, training::Training, data::AbstractData{2};
    callback = default_callback, outputdir,
    tempdir::AbstractString = Base.tempdir()
)

Train model using the training configuration on data.

The keyword arguments are as follows.

outputdir represents the path where StreamlinerCore saves model weights (can be local or remote).

tempdir represents a folder where StreamlinerCore will store temporary data during training.

callback(m, trace) will be called after every epoch.

The arguments of callback work as follows.

m is the instantiated neural network or machine,

trace is an object encoding additional information, i.e.,

stats (average of metrics computed so far),

metrics (functions used to compute stats), and

iteration.

StreamlinerCore.finetune — Function

finetune(
    result::Result, training::Training, data::AbstractData{2};
    resume::Bool, callback = default_callback,
    outputdir, tempdir::AbstractString = Base.tempdir()
)

Load model encoded in result and retrain it using the training configuration on data. Use resume = true to restart training where it left off. Same other keyword arguments as train.

StreamlinerCore.loadmodel — Function

loadmodel(model::Model, training::Training, data::AbstractData)

Load model encoded in model. The object data is required as the model can only be initialized once the data dimensions are known.

loadmodel(result::Union{Model, Result}, training::Training, data::AbstractData)

Load model encoded in result. The object data is required as the model can only be initialized once the data dimensions are known.

Warning

It is recommended to call has_weights beforehand. Only call loadmodel if has_weights(result) returns true.

StreamlinerCore.validate — Function

validate(result::Result, training::Training, data::AbstractData{1})

Load model encoded in result and validate it on data.

StreamlinerCore.evaluate — Function

evaluate(device_m, training::Training, data::AbstractData{1}, select::SymbolTuple = (:prediction,))

Evaluate model device_m on data using device and batchsize from training.

evaluate(
    result::Result, training::Training, data::AbstractData{1},
    select::SymbolTuple = (:prediction,)
)

Load model encoded in result and evaluate it on data.

StreamlinerCore.summarize — Function

summarize(io::IO, model::Model, training::Training, data::AbstractData)

Display summary information concerning model (structure and number of parameters) and data (number of batches and size of each batch).

Utilities

StreamlinerCore.has_weights — Function

has_weights(result::Result)

Return true if result is a successful training result, false otherwise.