Pipelines

Pipelines is a library designed to generate and evaluate data analysis pipelines.

Transformation interface

Pipelines.AbstractCard — Type

abstract type AbstractCard end

Abstract supertype to encompass all possible filters.

Current implementations:

RescaleCard,

SplitCard,

GLMCard.

Pipelines.train — Function

train(repo::Repository, card::AbstractCard, source; schema = nothing)

Return a trained model for a given card on a table table in the database repo.db.

Pipelines.evaluate — Function

evaluate(repo::Repository, card::AbstractCard, m, (source, target)::Pair; schema = nothing)

Replace table target in the database repo.db with the outcome of executing the card on the table source.

Here, m represents the result of train(repo, card, source; schema). See also train.

Pipelines.inputs — Function

outputs(c::AbstractCard)

Return the list of outputs for a given card.

Pipelines.outputs — Function

inputs(c::AbstractCard)

Return the list of inputs for a given card.

Pipeline computation

Pipelines.get_card — Function

get_card(d::AbstractDict)

Generate an AbstractCard based on a configuration dictionary.

Pipelines.evaluate — Method

evaluate(repo::Repository, cards::AbstractVector, table::AbstractString; schema = nothing)

Replace table in the database repo.db with the outcome of executing all the transformations in cards.

Cards

Pipelines.RescaleCard — Type

struct RescaleCard <: AbstractCard
    method::String
    by::Vector{String} = String[]
    columns::Vector{String}
    suffix::String = "rescaled"
end

Card to rescale of one or more columns according to a given method. The supported methods are

zscore,

maxabs,

minmax,

log,

logistic.

The resulting rescaled variable is added to the table under the name "$(originalname)_$(suffix)".

Pipelines.SplitCard — Type

struct SplitCard <: AbstractCard
    method::String
    order_by::Vector{String}
    by::Vector{String} = String[]
    output::String
    p::Float64 = NaN
    tiles::Vector{Int} = Int[]
end

Card to split the data into two groups according to a given method.

Currently supported methods are

tiles (requires tiles argument, e.g., tiles = [1, 1, 2, 1, 1, 2]),

percentile (requires p argument, e.g. p = 0.9).

Pipelines.GLMCard — Type

struct GLMCard <: AbstractCard
    predictors::Vector{Any} = Any[]
    target::String
    weights::Union{String, Nothing} = nothing
    distribution::String = "normal"
    link::Union{String, Nothing} = nothing
    link_params::Vector{Any} = Any[]
    suffix::String = "hat"
end

Run a GLM